some restructuring to move the `libgit2` modules into their own package `native`
