+++ tables/internal/git/native/blame.go
@@ -1,4 +1,4 @@
-package git
+package native
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/askgitdev/askgit/tables/internal/git"
 	"github.com/askgitdev/askgit/tables/services"
 	"github.com/augmentable-dev/vtab"
 	"github.com/go-git/go-git/v5/storage/filesystem"
@@ -47,7 +48,7 @@ func NewBlameModule(locator services.RepoLocator, ctx services.Context) sqlite.M
 
 		if repoPath == "" {
 			var err error
-			repoPath, err = getDefaultRepoFromCtx(ctx)
+			repoPath, err = git.GetDefaultRepoFromCtx(ctx)
 			if err != nil {
 				return nil, err
 			}
