+++ tables/internal/git/native/blame_test.go
@@ -1,4 +1,4 @@
-package git_test
+package native_test
 
 import (
 	"encoding/csv"
