+++ tables/internal/git/native/stats.go
@@ -1,10 +1,11 @@
-package git
+package native
 
 import (
 	"context"
 	"fmt"
 	"io"
 
+	"github.com/askgitdev/askgit/tables/internal/git"
 	"github.com/askgitdev/askgit/tables/services"
 	"github.com/augmentable-dev/vtab"
 	"github.com/go-git/go-git/v5/storage/filesystem"
@@ -41,7 +42,7 @@ func NewStatsModule(locator services.RepoLocator, ctx services.Context) sqlite.M
 
 		if repoPath == "" {
 			var err error
-			repoPath, err = getDefaultRepoFromCtx(ctx)
+			repoPath, err = git.GetDefaultRepoFromCtx(ctx)
 			if err != nil {
 				return nil, err
 			}
