+++ tables/internal/git/utils.go
@@ -21,9 +21,9 @@ func isRemoteBranch(ref plumbing.ReferenceName) bool {
 		plumbing.ReferenceName(strings.Replace(ref.String(), "remotes", "heads", 1)).IsBranch()
 }
 
-// getDefaultRepoFromCtx looks up the defaultRepoPath key in the supplied context and returns it if set,
+// GetDefaultRepoFromCtx looks up the defaultRepoPath key in the supplied context and returns it if set,
 // otherwise it returns the current working directory
-func getDefaultRepoFromCtx(ctx services.Context) (repoPath string, err error) {
+func GetDefaultRepoFromCtx(ctx services.Context) (repoPath string, err error) {
 	var ok bool
 	if repoPath, ok = ctx["defaultRepoPath"]; !ok || repoPath == "" {
 		if wd, err := os.Getwd(); err != nil {
